ml-service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
tensorflow>=2.18.0
pillow>=10.1.0
numpy>=1.26.0
python-multipart>=0.0.6
requests>=2.31.0
huggingface_hub>=0.24.0
tqdm>=4.67.0
filelock>=3.16.0
fsspec>=2023.5.0
python-dotenv>=1.0.1
onnx>=1.21.0
onnxruntime>=1.23.0
tf2onnx>=1.17.0
opencv-python>=4.13.0.0
sympy>=1.14.0
